Move UserButton module into lib
