+++ examples/button.rs
@@ -9,7 +9,7 @@ use stm32f3_discovery::prelude::*;
 use stm32f3_discovery::stm32;
 
 use stm32f3_discovery::leds::Led;
-use button::Button;
+use stm32f3_discovery::button::{Button, UserButton};
 
 #[entry]
 fn main() -> ! {
@@ -28,7 +28,7 @@ fn main() -> ! {
 
     // initialize user button
     let gpioa = device_periphs.GPIOA.split(&mut reset_and_clock_control.ahb);
-    let button = button::UserButton::new(gpioa.pa0);
+    let button = UserButton::new(gpioa.pa0);
 
     loop {
         delay.delay_ms(50u16);
@@ -37,38 +37,6 @@ fn main() -> ! {
             Ok(true) => status_led.on(),
             Ok(false) => status_led.off(),
             Err(_e) => panic!("Failed to read button state")
-        }
-    }
-}
-
-mod button {
-    use stm32f3xx_hal::gpio::gpioa::PA0;
-    use stm32f3xx_hal::gpio::{Input, Floating};
-    use stm32f3xx_hal::hal::digital::v2::InputPin;
-
-    type Error = ();
-
-    pub trait Button {
-        fn is_pressed(&self) -> Result<bool, Error>;
-    }
-
-    pub struct UserButton {
-        //Has an external pulldown & low pass filter.
-        pin: PA0<Input<Floating>>
-    }
-
-    impl UserButton {
-        pub fn new(pa0: PA0<Input<Floating>>) -> Self {
-            UserButton {
-                pin: pa0
-            }
-        }
-    }
-
-    //TODO: Make generic over active high buttons, i.e. Floating (w/ external pull down) + PullDown
-    impl Button for UserButton {
-        fn is_pressed(&self) -> Result<bool, Error> {
-            self.pin.is_high()
         }
     }
 }
